# swagger.yaml
openapi: 3.0.0
info:
  title: ecart-testing
  contact: {}
  version: '1.0'
servers:
- url: https://temp-ecom-api.onrender.com/api
  variables: {}
paths:
  /product/all-products/:id:
    get:
      tags:
      - Public-Access
      summary: get single product
      operationId: getsingleproduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /auth/sign-up:
    post:
      tags:
      - Public-Access
      summary: register
      operationId: register
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/registerrequest'
              - example:
                  name: ashuto123
                  email: ashu123@gmail.com
                  password: '123'
            example:
              name: ashuto123
              email: ashu123@gmail.com
              password: '123'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /auth/sign-in:
    post:
      tags:
      - Public-Access
      summary: log-in user
      operationId: log-inuser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/log-inuserrequest'
              - example:
                  email: baba55442@gmail.com
                  password: '12345'
            example:
              email: baba55442@gmail.com
              password: '12345'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /product/all-products:
    get:
      tags:
      - Public-Access
      summary: get all product
      operationId: getallproduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/update-user/{ID}:
    post:
      tags:
      - Auth
      summary: update-user
      description: >-
        http://localhost:3000/api/user/update-user/:ID

        {name and password only}
      operationId: update-user
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/update-user-request'
              - example:
                  name: 5442@gmail
            example:
              name: 5442@gmail
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /user/update-user/{id}:
    delete:
      tags:
      - Auth
      summary: Delete user
      operationId: Deleteuser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
    registerrequest:
      title: registerrequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: ashuto123
        email: ashu123@gmail.com
        password: '123'
    log-inuserrequest:
      title: log-inuserrequest
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: baba55442@gmail.com
        password: '12345'
    update-user-request:
      title: update-user-request
      required:
      - name
      type: object
      properties:
        name:
          type: string
      example:
        name: 5442@gmail
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
tags:

- name: Public-Access
  description: ''
- name: Auth
  description: